requirements.txt
PyQt5>=5.15.0
pyinstaller>=4.0
PyYAML>=5.0
pywin32>=227
